Deaths are processed by the GameMode and Notifications are sent from the GameState
Notifications are done using UE4 Events (kinda messy).
Yet to implement the Notification Feed (just using Print String for now).
GameMode can detect who the winner of the round is and how many players are still alive.
Added an IsDead field to PlayerState so we know when the player is dead.